pyproject.toml
[project]
name = "market-insight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "langchain>=1.1.3",
    "langchain-openai>=1.1.3",
    "langfuse>=3.10.6",
    "langgraph>=1.0.5",
    "openpyxl>=3.1.5",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
]
